Abstracted Slider from TimeSlider
